Fix no radical sign in $\sqrt$
